在打车界面通过左侧的slidingMenu切换打车界面不同 CallCarMenuDetailPager Ps://今天周六从老家回来，天气很好
