+++ app/src/main/java/com/tantuo/didicar/pager/GovaffairPager.java
@@ -27,7 +27,7 @@ public class GovaffairPager extends BasePager {
         super.initData();
         LogUtil.i("合规数据初始化..");
         //1.设置标题
-        tv_title.setText("主页面");
+        tv_title.setText("合规界面");
         //2.联网请求，得到数据，创建视图
         TextView textView = new TextView(context);
         textView.setGravity(Gravity.CENTER);
@@ -36,7 +36,7 @@ public class GovaffairPager extends BasePager {
         //3.把子视图添加到BasePager的FrameLayout中
         fl_content.addView(textView);
         //4.绑定数据
-        textView.setText("主页面内容");
+        textView.setText("合规数据初始化");
 
     }
 }
