+++ app/src/main/java/com/tantuo/didicar/pager/SettingPager.java
@@ -24,7 +24,7 @@ public class SettingPager extends BasePager {
         super.initData();
         LogUtil.i("设置页面初始化..");
         //1.设置标题
-        tv_title.setText("666666");
+        tv_title.setText("设置界面");
         //2.联网请求，得到数据，创建视图
         TextView textView = new TextView(context);
         textView.setGravity(Gravity.CENTER);
@@ -33,7 +33,7 @@ public class SettingPager extends BasePager {
         //3.把子视图添加到BasePager的FrameLayout中
         fl_content.addView(textView);
         //4.绑定数据
-        textView.setText("45rttttrr");
+        textView.setText("设置界面数据");
 
     }
 }
