+++ app/src/main/res/layout/titlebar.xml
@@ -21,7 +21,7 @@
         android:layout_marginLeft="8dp"
         android:background="@android:color/transparent"
         android:src="@drawable/img_menu"
-        android:visibility="visible" />
+        android:visibility="gone" />
 
     <ImageButton
         android:id="@+id/ib_back"
